backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
apify-client
python-dotenv
google-genai
yt-dlp
ffmpeg-python